+++ CUDATools.h
@@ -127,9 +127,9 @@ public:
     NEListToArray(mesh->NEList);
 
     // and copy everything to the device
-    copyArrayToDevice<double>(&mesh->coords[0], CUDA_coords, NNodes * ndims);
-    copyArrayToDevice<double>(&mesh->metric[0], CUDA_metric, NNodes * nloc); //TODO is thhis right?
-    copyArrayToDevice<double>(&mesh->normals[0], CUDA_normals, NNodes * ndims);
+    copyArrayToDevice<float>(&mesh->coords[0], CUDA_coords, NNodes * ndims);
+    copyArrayToDevice<float>(&mesh->metric[0], CUDA_metric, NNodes * nloc); //TODO is thhis right?
+    copyArrayToDevice<float>(&mesh->normals[0], CUDA_normals, NNodes * ndims);
     copyArrayToDevice<size_t>(&mesh->ENList[0], CUDA_ENList, NElements * nloc);
     copyArrayToDevice<size_t>(NNListArray, CUDA_NNListArray, NNListArray_size);
     copyArrayToDevice<size_t>(NNListIndex, CUDA_NNListIndex, NNodes+1);
@@ -161,22 +161,22 @@ public:
 
   void copyCoordinatesToDevice(Mesh* mesh)
   {
-    copyArrayToDeviceNoAlloc<double>(&mesh->coords[0], CUDA_coords, NNodes * ndims);
+    copyArrayToDeviceNoAlloc<float>(&mesh->coords[0], CUDA_coords, NNodes * ndims);
   }
 
   void copyMetricToDevice(Mesh* mesh)
   {
-    copyArrayToDeviceNoAlloc<double>(&mesh->metric[0], CUDA_metric, NNodes * nloc);
+    copyArrayToDeviceNoAlloc<float>(&mesh->metric[0], CUDA_metric, NNodes * nloc);
   }
 
   void copyCoordinatesFromDevice(Mesh* mesh)
   {
-    copyArrayFromDevice<double>(&mesh->coords[0], CUDA_coords, NNodes * ndims);
+    copyArrayFromDevice<float>(&mesh->coords[0], CUDA_coords, NNodes * ndims);
   }
 
   void copyMetricFromDevice(Mesh* mesh)
   {
-    copyArrayFromDevice<double>(&mesh->metric[0], CUDA_metric, NNodes * nloc);
+    copyArrayFromDevice<float>(&mesh->metric[0], CUDA_metric, NNodes * nloc);
   }
 
   void freeResources()
